renamed a taskCard Folder to taskComponents
